[GLideN64-wtl] Move util.h to util/util.h
