+++ projects/msvc/GLideNUI-wtl.vcxproj.filters
@@ -72,7 +72,7 @@
     <ClCompile Include="..\..\src\GLideNUI-wtl\Settings.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\GLideNUI-wtl\util.cpp">
+    <ClCompile Include="..\..\src\GLideNUI-wtl\util\util.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\GLideNUI-wtl\wtl-BitmapPicture.cpp">
@@ -149,7 +149,7 @@
     <ClInclude Include="..\..\src\GLideNUI-wtl\Settings.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\GLideNUI-wtl\util.h">
+    <ClInclude Include="..\..\src\GLideNUI-wtl\util\util.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\GLideNUI-wtl\wtl.h">
@@ -247,9 +247,6 @@
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <Image Include="..\..\src\GLideNUI-wtl\Icon-Original.ico">
-      <Filter>Resource Files</Filter>
-    </Image>
     <Image Include="..\..\src\GLideNUI\Resources\BottomLeft.ico">
       <Filter>Resource Files</Filter>
     </Image>
