+++ src/GLideNUI-wtl/Language.cpp
@@ -1,5 +1,5 @@
 #include "Language.h"
-#include "util.h"
+#include "util/util.h"
 #include <windows.h>
 #include <algorithm>
 #include <map>
