+++ src/GLideNUI-wtl/config-emulation.cpp
@@ -1,5 +1,5 @@
 #include "config-emulation.h"
-#include "util.h"
+#include "util/util.h"
 #include "../Config.h"
 #include "resource.h"
 #include "ConfigDlg.h"
