+++ src/GLideNUI-wtl/config-osd.cpp
@@ -1,6 +1,6 @@
 #include "config-osd.h"
 #include "resource.h"
-#include "util.h"
+#include "util/util.h"
 #include "../Config.h"
 #include "FontInfo.h"
 #include "Language.h"
