+++ src/GLideNUI-wtl/config-overscan.cpp
@@ -1,5 +1,5 @@
 #include "config-overscan.h"
-#include "util.h"
+#include "util/util.h"
 #include "resource.h"
 
 COverScanTab::COverScanTab()
