+++ src/GLideNUI-wtl/util/util.cpp
