+++ src/GLideNUI-wtl/util/util.h
